extract cache_storage.rs
